Create ToolInstallCommand as redirector
Since global tool install and local tools install don't have too much to share
the common part is individual module already
